requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
psycopg2-binary==2.9.10
redis==5.2.1
elasticsearch==8.18.1
python-dotenv==1.1.0
httpx==0.28.1
